Refactoring: extracted SpotlightClientCommandlineDecoder and TkInterface classes
